don't propagate Guard.listener.ignore_paths to rsync excludes file
guard/guard@306bd4de deprecated the 'ignore_paths' directive
previously available in the Guardfile DSL, in favour of new 'ignore'
and 'filter' directives which accept regexp parameters:

    https://github.com/guard/guard/commit/306bd4de

This means it is no longer practical to propagate the values to
rsync's exclude file, because there is no easy way to convert from
regexps to exclude patterns.

Even if there was, it would not be correct to do it, because 'ignore'
regexps are intended to control the guard listener, which is
independent to the expected behaviour of the rsync invocation.

For example, the user might want certain files only to be rsynced via
the 'run_all' action, and not when they individually change.  This
could be achieved by specifying them in an 'ignore' directive, but not
in guard-rsync's ':excludes' option.

Or the user might choose to compile .sass files to CSS via a
run_on_changes_begin callback:

    https://github.com/guard/guard/wiki/Hooks-and-callbacks

and then rsync only the resulting .css files to the target.  In this
case, the 'ignore' directive could ignore changes to .css files (or
the 'filter' directive could ignore all changes except those to .sass
files) and then the rsync would either exclude .sass files or only
include .css files.