simplify structure of :excludes option values
The :excludes option expected a Hash whose keys were Regexps.  This
had a number of issues:

  1. It required iterating over the whole source tree to find which
     files matched; this could be costly if the tree is large.

  2. The values were expected to be nil, false, or Procs which would
     transform a matching file path into another file path to be
     excluded.  But this syntax is ugly when you just want to exclude
     a single thing, e.g.

       :excludes => { '.git' => false }

     and didn't actually buy anything useful, since the exclusions
     caused by calling the Proc could be equivalently achieved via
     another Regexp.

  3. It's counter-intuitive since most users already know rsync
     accepts *lists* of exclude patterns, not hashes.

Therefore we simply switch the exclude list to be an Array of native
rsync exclude patterns.  Later we can add back support for Regexps in
case the convenience afforded to the user outweighs the
Dir.glob('**/*') performance hit.